+++ lib/src/collections/collections.dart
@@ -2,5 +2,4 @@
 /// SPDX-License-Identifier: BSD-3-Clause
 
 export 'duplicate_detection_set.dart';
-export 'synchronous_propagator.dart';
 export 'traverseable_collection.dart';
