+++ lib/src/collections/duplicate_detection_set.dart
