+++ lib/src/collections/traverseable_collection.dart
