+++ lib/src/utilities/synchronous_propagator.dart
